Using function in compilation stage
